+++ packages/interactive_media_ads/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.2.4+2
+
+* Bumps gradle from 8.9.0 to 8.11.1.
+
 ## 0.2.4+1
 
 * Adds internal wrapper for Android native `ImaSdkSettings`.